Initialize CM instance outside of first request Resolves: https://www.notion.so/lyrasis/Initialize-CM-instance-outside-of-first-request-d024b0b64174442d8bd2c2171a2d4aeb?pvs=4
